forgot to remove the redundant (float) cast, as specified by reviewer
Canonical link: https://commits.webkit.org/9876@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@11804 268f45cc-cd09-0410-ab3c-d52691b4dbfc

+++ WebKit/WebCoreSupport.subproj/WebTextRenderer.m
@@ -1381,7 +1381,7 @@ static void createATSULayoutParameters(ATSULayoutParameters *params, WebTextRend
             if (isSpace(run->characters[k]))
                 numSpaces++;
 
-        params->padPerSpace = ceilf((float)style->padding / numSpaces);
+        params->padPerSpace = ceilf(style->padding / numSpaces);
     } else
         params->padPerSpace = 0;
 }
